--- a/MatrixSDK/MXSession.m
+++ b/MatrixSDK/MXSession.m
@@ -302,6 +302,7 @@ typedef void (^MXOnResumeDone)(void);
     }
 
     NSDate *startDate = [NSDate date];
+    MXTaskProfile *taskProfile = [MXSDKOptions.sharedInstance.profiler startMeasuringTaskWithName:kMXAnalyticsStartupMountData category:kMXAnalyticsStartupCategory];
 
     MXWeakify(self);
     [_store openWithCredentials:matrixRestClient.credentials onComplete:^{
@@ -365,14 +366,12 @@ typedef void (^MXOnResumeDone)(void);
                     [self loadRoom:roomId];
                 }
 
-                NSLog(@"[MXSession] Built %lu MXRooms in %.0fms", (unsigned long)self->rooms.allKeys.count, [[NSDate date] timeIntervalSinceDate:startDate3] * 1000);
+                NSLog(@"[MXSession] Built %lu MXRooms in %.0fms", (unsigned long)self->rooms.count, [[NSDate date] timeIntervalSinceDate:startDate3] * 1000);
 
-                NSTimeInterval duration = [[NSDate date] timeIntervalSinceDate:startDate];
-                NSLog(@"[MXSession] Total time to mount SDK data from MXStore: %.0fms", duration * 1000);
-
-                [[MXSDKOptions sharedInstance].analyticsDelegate trackDuration:duration
-                                                                      category:kMXAnalyticsStartupCategory
-                                                                          name:kMXAnalyticsStartupMountData];
+                taskProfile.units = self->rooms.count;
+                [MXSDKOptions.sharedInstance.profiler stopMeasuringTaskWithProfile:taskProfile];
+                
+                NSLog(@"[MXSession] Total time to mount SDK data from MXStore: %.0fms", taskProfile.duration * 1000);
                 
                 [self setState:MXSessionStateStoreDataReady];
 
@@ -950,6 +949,14 @@ typedef void (^MXOnResumeDone)(void);
                         setPresence:(NSString*)setPresence
 {
     NSDate *startDate = [NSDate date];
+    
+    MXTaskProfile *syncTaskProfile;
+    if (!self->firstSyncDone)
+    {
+        BOOL isInitialSync = !self.isEventStreamInitialised;
+        syncTaskProfile = [MXSDKOptions.sharedInstance.profiler startMeasuringTaskWithName:isInitialSync ? kMXAnalyticsStartupInititialSync : kMXAnalyticsStartupIncrementalSync
+                                                        category:kMXAnalyticsStartupCategory];
+    }
 
     // Determine if we are catching up
     _catchingUp = (0 == serverTimeout);
@@ -977,14 +984,12 @@ typedef void (^MXOnResumeDone)(void);
         BOOL isInitialSync = !self.isEventStreamInitialised;
 
         BOOL wasfirstSync = NO;
-        if (!self->firstSyncDone)
+        if (!self->firstSyncDone && syncTaskProfile)
         {
             wasfirstSync = YES;
             self->firstSyncDone = YES;
             
-            [[MXSDKOptions sharedInstance].analyticsDelegate trackDuration:duration
-                                                                  category:kMXAnalyticsStartupCategory
-                                                                      name:isInitialSync ? kMXAnalyticsStartupInititialSync : kMXAnalyticsStartupIncrementalSync];
+            [MXSDKOptions.sharedInstance.profiler stopMeasuringTaskWithProfile:syncTaskProfile];
         }
 
         // Handle the to device events before the room ones