--- a/MatrixSDK/Utils/Profiling/MXBaseProfiler.h
+++ b/MatrixSDK/Utils/Profiling/MXBaseProfiler.h
@@ -1,0 +1,33 @@
+// 
+// Copyright 2020 The Matrix.org Foundation C.I.C
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+#import <Foundation/Foundation.h>
+
+#import "MXProfiler.h"
+#import "MXAnalyticsDelegate.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+/**
+ A profiler that sends data to analytics.
+ */
+@interface MXBaseProfiler : NSObject <MXProfiler>
+
+@property (nonatomic, weak, nullable) id<MXAnalyticsDelegate> analytics;
+
+@end
+
+NS_ASSUME_NONNULL_END