--- a/MatrixSDK/Utils/Profiling/MXBaseProfiler.m
+++ b/MatrixSDK/Utils/Profiling/MXBaseProfiler.m
@@ -54,7 +54,7 @@
 {
     [taskProfile markAsCompleted];
     
-    NSLog(@"[MXBaseProfiler] Task %@ - %@ for %@ units completed in  %.3fms%@",
+    NSLog(@"[MXBaseProfiler] Task %@ - %@ for %@ units completed in %.3fms%@",
           taskProfile.category,
           taskProfile.name,
           @(taskProfile.units),