--- a/MatrixSDK/Utils/Profiling/MXBaseProfiler.m
+++ b/MatrixSDK/Utils/Profiling/MXBaseProfiler.m
@@ -74,6 +74,15 @@
     }
 }
 
+- (void)cancelTaskProfile:(MXTaskProfile *)taskProfile;
+{
+    @synchronized (taskProfiles)
+    {
+        [taskProfiles removeObject:taskProfile];
+    }
+}
+
+
 - (void)pause
 {
     @synchronized (taskProfiles)