… CHANGES.rst
… MatrixSDK/Utils/Profiling/MXTaskProfile_Private.h