The default implementation, `MXBaseProfiler`, reports duration in logs and to the `MXAnalytics` interface.
Logs look like:
```
[MXBaseProfiler] Task startup - storePreload for 578 units completed in 1682.477ms
[MXBaseProfiler] Task startup - mountData for 578 units completed in 2267.157ms
[MXBaseProfiler] Task startup - launchScreen for 1 units completed in 2282.014ms
...
[MXBaseProfiler] Task startup - incrementalSync for 53 units completed in 481.103ms
[MXBaseProfiler] Task notifications - timelineDisplay for 1 units completed in 652.003ms
```
 
The app can call `MXProfiler.pause()` when it goes in background. This avoids to measure and to report bad durations.